Created a subscribe to newsletter molecule and migrated the code from newsletter content there.
