+++ src/stories/molecules/subscribe-to-newsletter.stories.tsx
@@ -1,0 +1,9 @@
+import SubscribeToNewsletter from "components/molecules/subscribe-to-newsletter/subscribe-to-newsletter";
+
+const storyConfig = {
+  title: "Design System/Molecules/Subscribe to Newsletter"
+};
+
+export default storyConfig;
+
+export const SubscribeToNewsletterStory = () => <SubscribeToNewsletter />;
